Add environment and target
