--- packages/browser-data-collector/src/api/calypso-version.ts
@@ -1,3 +1,0 @@
-export const getCalypsoVersion = (): string => {
-	return window?.COMMIT_SHA;
-};
