--- packages/browser-data-collector/src/collectors/calypso-version.ts
@@ -1,9 +1,0 @@
-/**
- * Internal dependencies
- */
-import { getCalypsoVersion } from '../api/calypso-version';
-
-export const collector: Collector = ( report ) => {
-	report.data.set( 'version', getCalypsoVersion() );
-	return report;
-};
